+++ tests/dummy/app/controllers/besluit-sample.ts
@@ -213,6 +213,20 @@ export default class BesluitSampleController extends Controller {
           return new Set<NodeType>([schema.nodes.motivering]);
         },
       },
+      roadsignRegulation: {
+        endpoint: 'https://dev.roadsigns.lblod.info/sparql',
+        imageBaseUrl: 'https://register.mobiliteit.vlaanderen.be/',
+      },
+      besluitType: {
+        endpoint: 'https://centrale-vindplaats.lblod.info/sparql',
+      },
+      templateVariable: {
+        endpoint: 'https://dev.roadsigns.lblod.info/sparql',
+        zonalLocationCodelistUri:
+          'http://lblod.data.gift/concept-schemes/62331E6900730AE7B99DF7EF',
+        nonZonalLocationCodelistUri:
+          'http://lblod.data.gift/concept-schemes/62331FDD00730AE7B99DF7F2',
+      },
       link: {
         interactive: true,
       },